Drop propagatedBuildInputs in favor of explicitly requiring rustc
This will make it easier for consumers to use the version of rustc they
want to use with fewer surprises. E.g. even if they don't use
`overrideToolchain` but provide a specific version of `rustc` in their
nativeBuildInputs, theirs should take precedence (earlier entries show up
earlier in the path) which might be harder to reason about if its coming
in as a propagated build input